Removed doubled update functions
